requirements.txt
# ==================== Web框架 ====================
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
starlette>=0.38.0

# ==================== 数据库 ====================
sqlalchemy>=2.0.0
alembic>=1.13.0

# ==================== 认证与安全 ====================
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-jose>=3.3.0
python-multipart>=0.0.6
cryptography>=41.0.0

# ==================== HTTP客户端 ====================
httpx>=0.26.0
requests>=2.31.0
aiohttp>=3.9.0

# ==================== AI服务 ====================
# 阿里云通义千问 API
dashscope>=1.14.1

# Google Gemini API
google-generativeai>=0.8.0
google-api-python-client>=2.0.0
google-auth>=2.23.0

# ==================== 数据处理 ====================
pandas>=2.0.0
numpy>=1.24.0

# ==================== 图像处理 ====================
Pillow>=10.0.0
opencv-python>=4.8.0

# ==================== 数据分析与可视化（可选） ====================
matplotlib>=3.7.0
seaborn>=0.13.0
scipy>=1.10.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# ==================== 数据验证 ====================
pydantic>=2.5.0
pydantic-settings>=2.0.0

# ==================== 工具 ====================
python-dotenv>=1.0.0
python-dateutil>=2.8.0
pytz>=2023.3
tqdm>=4.66.0
